+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/api/NetIdServiceListener.kt
@@ -27,4 +27,6 @@ interface NetIdServiceListener {
     fun onPermissionFetchFinished(permissions: Permissions)
 
     fun onPermissionUpdateFinished()
+
+    fun onTransmittedInvalidToken()
 }