+++ CHANGELOG.md
@@ -9,3 +9,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Inital commit
 - Permission management webservices and API functions
 - AuthorizationView screen design alignment
+- API function to transmit a id token